updated for version 7.3.037 Problem:    Compiler warnings for loss of data. (Mike Williams) Solution:   Add type casts.

+++ src/getchar.c
@@ -3922,7 +3922,7 @@ showmap(mp, local)
     if (mapchars != NULL)
     {
 	msg_puts(mapchars);
-	len = STRLEN(mapchars);
+	len = (int)STRLEN(mapchars);
 	vim_free(mapchars);
     }
